+++ vignettes/custom.R
@@ -9,6 +9,7 @@ PlotTree <- function(tree, ...) {
   par(oPar)
 }
 
+
 ## ----tci-setup----------------------------------------------------------------
 tree <- PectinateTree(8)
 PlotTree(tree)
@@ -20,14 +21,15 @@ TCIScore <- function(parent, child, dataset) {
 
 TCIScore(tree$edge[, 1], tree$edge[, 2], NA)
 
+
 ## ----tci-search---------------------------------------------------------------
 result <- TreeSearch(tree, dataset = EmptyPhyDat(tree),
-                     InitializeData = DoNothing, CleanUpData = DoNothing,
                      TreeScorer = TCIScore,
-                     maxIter = 50L, maxHits = 10L, 
+                     maxIter = 50L, maxHits = 10L,
                      verbosity = 1L)
 
 PlotTree(result)
+
 
 ## ----cid----------------------------------------------------------------------
 startTree <- BalancedTree(8)
@@ -41,9 +43,8 @@ DistanceScore <- function(parent, child, dataset) {
 }
 
 result <- TreeSearch(RandomTree(8, root = TRUE), dataset = EmptyPhyDat(tree),
-                     InitializeData = DoNothing, CleanUpData = DoNothing,
                      TreeScorer = DistanceScore,
-                     maxIter = 50L, maxHits = 10L, 
+                     maxIter = 50L, maxHits = 10L,
                      verbosity = 1L)
 
 par(mfrow = c(1, 2))
@@ -51,89 +52,48 @@ PlotTree(startTree)
 PlotTree(result)
 
 
-## ----iw-setup-----------------------------------------------------------------
-IWInitMorphy <- function (dataset) {
-  attr(dataset, "morphyObjs") <- 
-    lapply(PhyToString(dataset, byTaxon = FALSE, useIndex = FALSE, 
-                       concatenate = FALSE), 
-           SingleCharMorphy)
-  
-  # Return:
-  dataset
-}
 
-## ----iw-destroy---------------------------------------------------------------
-IWDestroyMorphy <- function (dataset) {
-  vapply(attr(dataset, "morphyObjs"), UnloadMorphy, integer(1))
-}
+## ----iw-score, message = FALSE------------------------------------------------
+data("inapplicable.datasets")
+dataset <- congreveLamsdellMatrices[[42]]
 
-## ----iw-score-----------------------------------------------------------------
-IWScoreMorphy <- function (parent, child, dataset, concavity = 10L, 
-                           minLength = attr(dataset, "min.length"), ...) {
-  steps <- vapply(attr(dataset, "morphyObjs"), MorphyLength,
-                  parent = parent, child = child, integer(1))
-  homoplasies <- steps - minLength
-  fit <- homoplasies / (homoplasies + concavity)
+IWScore <- function (parent, child, dataset, concavity = 10,
+                     minLength = MinimumLength(dataset, compress = TRUE)) {
+  tree <- structure(list(edge = cbind(parent, child),
+                        tip.label = names(dataset),
+                        Nnode = length(dataset) - 1L), class = "phylo")
+  homoplasy <- CharacterLength(tree, dataset, compress = TRUE) - minLength
+  fit <- homoplasy / (homoplasy + concavity)
   # Return:
   sum(fit * attr(dataset, "weight"))
 }
 
-## ----iw-search, message = FALSE-----------------------------------------------
-data("inapplicable.datasets")
-dataset <- congreveLamsdellMatrices[[42]]
-
-# Populate `min.length` attribute
-dataset <- PrepareDataIW(dataset)
-iwTree <- TreeSearch(NJTree(dataset), dataset,
-                     InitializeData = IWInitMorphy,
-                     CleanUpData = IWDestroyMorphy,
-                     TreeScorer = IWScoreMorphy,
-                     concavity = 10, # Will be sent to TreeScorer
-                     verbosity = 1)
+iwTree <- TreeSearch(NJTree(dataset), dataset, TreeScorer = IWScore,
+                     maxIter = 50L, maxHits = 10L, verbosity = 1)
 
 
 ## ----iw-bootstrap-------------------------------------------------------------
-IWBootstrap <- function (edgeList, dataset, concavity = 10L, EdgeSwapper = NNISwap, 
+IWBootstrap <- function (edgeList, dataset, EdgeSwapper = NNISwap,
                          maxIter, maxHits, verbosity = 1L, ...) {
-  att <- attributes(dataset)
-  startWeights <- att[["weight"]]
-  
-  # Decompress phyDat object so each character is listed once
-  eachChar <- seq_along(startWeights)
-  deindexedChars <- rep.int(eachChar, startWeights)
-  
-  # Resample characters
-  resampling <- tabulate(sample(deindexedChars, replace = TRUE), length(startWeights))
-  sampled <- resampling != 0
-  sampledData <- lapply(dataset, function (x) x[sampled])
-  sampledAtt <- att
-  sampledAtt[["index"]] <- rep.int(seq_len(sum(sampled)), resampling[sampled])
-  sampledAtt[["weight"]] <- resampling[sampled]
-  sampledAtt[["nr"]] <- length(sampledAtt[["weight"]])
-  sampledAtt[["min.length"]] <- minLength <- att[["min.length"]][sampled]
-  sampledAtt[["morphyObjs"]] <- att[["morphyObjs"]][sampled]
-  attributes(sampledData) <- sampledAtt
-  
-  # Search using resampled dataset
-  res <- EdgeListSearch(edgeList[1:2], sampledData, TreeScorer = IWScoreMorphy,
-                        concavity = concavity, minLength = minLength,
-                        EdgeSwapper = EdgeSwapper, 
+  startWeights <- attr(dataset, "weight")
+  resampling <- tabulate(sample(rep.int(seq_along(startWeights), startWeights),
+                                replace = TRUE), length(startWeights))
+  # R copy-on-modify: the caller's `dataset` is unchanged.
+  attr(dataset, "weight") <- as.integer(resampling)
+
+  res <- EdgeListSearch(edgeList[1:2], dataset, TreeScorer = IWScore,
+                        EdgeSwapper = EdgeSwapper,
                         maxIter = maxIter, maxHits = maxHits,
-                        verbosity = verbosity - 1L)
-  
+                        verbosity = verbosity - 1L, ...)
+  # Return:
   res[1:2]
 }
 
 
 ## ----iw-ratchet, message = FALSE----------------------------------------------
-ratchetTree <- Ratchet(tree = iwTree, dataset = dataset,
-                       concavity = 10,
-                       InitializeData = IWInitMorphy, 
-                       CleanUpData = IWDestroyMorphy,
-                       TreeScorer = IWScoreMorphy,
+ratchetTree <- Ratchet(iwTree, dataset, TreeScorer = IWScore,
                        Bootstrapper = IWBootstrap,
                        ratchIter = 2, ratchHits = 2,
                        searchIter = 20, searchHits = 10,
                        verbosity = 2)
-